[#1108] - Mueve server routes a nuevo archivo
